Finished the project

+++ puzzleSolver/src/puzzleSolver/BoardComparator.java
@@ -7,6 +7,6 @@ public class BoardComparator implements Comparator<PuzzleBoard> {
 
     @Override
     public int compare(PuzzleBoard o1, PuzzleBoard o2) {
-        return o1.cost - o2.cost;
+        return (int)(o1.cost - o2.cost);
     }
 }
